requirements.txt
flask
numpy
tensorflow
gunicorn
pillow
streamlit
